+++ CHANGELOG.md
@@ -38,6 +38,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/), and t
 - Change useForm hook
   - Add support for dropdown fields
   - Add onBlur property to checkbox fields
+  - Add hasChanges function
+  - Update state when fieldsInfo argument changes
 
 ### 📆 Deprecated